Replaces EqualFunc with template functions.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@82 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ math/src/mathfunctions.cpp
@@ -17,4 +17,4 @@ const float vislib::math::FLOAT_EPSILON = 1e-5f;
 /*
  * vislib::math::DOUBLE_EPSILON
  */
-const float vislib::math::DOUBLE_EPSILON = 1e-7f;
+const double vislib::math::DOUBLE_EPSILON = 1e-7f;
